REFACTOR : DELETE- ScoreService <- StudentScoreRead 이식, ScoreMenu 수정
